gui funcionando
