+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".moredetails.presentation.OtherInfoView"
+        <activity android:name=".moredetails.presentation.view.OtherInfoView"
            android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
