+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/presentation/presenter/OtherInfoUiState.kt
@@ -1,4 +1,4 @@
-package ayds.lisboa.songinfo.moredetails.presentation
+package ayds.lisboa.songinfo.moredetails.presentation.presenter
 
 import ayds.lisboa.songinfo.moredetails.domain.entities.Card
