+++ app/src/test/java/ayds/lisboa/songinfo/moredetails/presenter/OtherInfoPresenterTest.kt
@@ -3,6 +3,9 @@ package ayds.lisboa.songinfo.moredetails.presenter
 import ayds.lisboa.songinfo.moredetails.domain.entities.Artist
 import ayds.lisboa.songinfo.moredetails.domain.repository.CardRepository
 import ayds.lisboa.songinfo.moredetails.presentation.*
+import ayds.lisboa.songinfo.moredetails.presentation.presenter.CardResolver
+import ayds.lisboa.songinfo.moredetails.presentation.presenter.OtherInfoPresenterImpl
+import ayds.lisboa.songinfo.moredetails.presentation.presenter.OtherInfoUiState
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
